Move stubs into separate package
